Move world stuff into class & game window size
It does not know anything about the window being shows or anything.
It just always renders everything in the world.
Simple as that.

The stuff that is rendered in the world is then cropped and positioned correctly outside of it, to make it seem like the world is being seen through the window.